Re #307 (code cleanup) Fixed MSVC warnings
git-svn-id: https://svn.code.sf.net/p/speed-dreams/code/trunk@3284 30fe4595-0a0c-4342-8851-515496e4dcbd

Former-commit-id: fa2607c6d87b2138184914f05721497eb20efbf1
Former-commit-id: 17721df15ca01734550eba1c7ae4fd8de14eede0

+++ src/libs/tgf/tgf.cpp
@@ -950,5 +950,5 @@ int GfNearestPow2 (int x)
 void
 GfSleep(double seconds)
 {
-  SDL_Delay(seconds*1000); // ms.
+  SDL_Delay((Uint32)(seconds*1000)); // ms.
 }
